requirements.txt
Flask==2.2.0
Flask-SQLAlchemy
wheel
pyjwt
datetime
uuid
pytest==7.1.2
coverage==6.4.2
python-dotenv
alembic
mysqlclient
flask-marshmallow
marshmallow-sqlalchemy
cerberus
Flask-Mail
celery
redis
requests